refactor: 💡 multi version architecture for count tM
