+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/3.0.0.ts
@@ -5,27 +5,22 @@ import {
   CountTransferManagerModifyHolderCountEventArgs_3_0_0,
   CountTransferManagerPauseEventArgs_3_0_0,
   CountTransferManagerUnpauseEventArgs_3_0_0,
-  Web3Wrapper,
   LogWithDecodedArgs,
-  BigNumber,
-  PolyResponse,
+  Web3Wrapper,
 } from '@polymathnetwork/abi-wrappers';
 import { schemas } from '@0x/json-schemas';
-import assert from '../../../utils/assert';
-import ModuleWrapper from '../module_wrapper';
-import ContractFactory from '../../../factories/contractFactory';
+import assert from '../../../../utils/assert';
 import {
-  TxParams,
-  GetLogsAsyncParams,
+  ContractVersion,
   SubscribeAsyncParams,
-  EventCallback,
+  GetLogsAsyncParams,
   Subscribe,
   GetLogs,
-  Perm,
-  ErrorCode,
-  TransferResult,
-} from '../../../types';
-import { numberToBigNumber, parseTransferResult, valueToWei } from '../../../utils/convert';
+  EventCallback,
+} from '../../../../types';
+import ContractFactory from '../../../../factories/contractFactory';
+
+import CountTransferManagerCommon from './common';
 
 interface ModifyHolderCountSubscribeAsyncParams extends SubscribeAsyncParams {
   eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
@@ -68,47 +63,13 @@ interface GetCountTransferManagerLogsAsyncParams extends GetLogs {
   (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerUnpauseEventArgs_3_0_0>[]>;
 }
 
-export namespace CountTransferManagerTransactionParams {
-  export interface ChangeHolderCount extends ChangeHolderCountParams {}
-}
-
-/**
- * @param from Address of the sender
- * @param to Address of the receiver
- * @param amount Amount to send
- * @param data Data value
- */
-interface VerifyTransferParams {
-  from: string;
-  to: string;
-  amount: BigNumber;
-  data: string;
-}
-
-/**
- * @param maxHolderCount is the new maximum amount of token holders
- */
-interface ChangeHolderCountParams extends TxParams {
-  maxHolderCount: number;
-}
-
-/**
- * @param transferResult
- * @param address
- */
-interface VerifyTransfer {
-  transferResult: TransferResult;
-  address: string;
-}
-
-/**
- * This class includes the functionality related to interacting with the Count Transfer Manager contract.
- */
-export default class CountTransferManagerWrapper extends ModuleWrapper {
+export class CountTransferManager_3_0_0 extends CountTransferManagerCommon {
   public contract: Promise<CountTransferManagerContract_3_0_0>;
 
+  public contractVersion = ContractVersion.V3_0_0;
+
   /**
-   * Instantiate CountTransferManagerWrapper
+   * Instantiate CountTransferManager_3_0_0
    * @param web3Wrapper Web3Wrapper instance to use
    * @param contract
    * @param contractFactory
@@ -121,83 +82,6 @@ export default class CountTransferManagerWrapper extends ModuleWrapper {
     super(web3Wrapper, contract, contractFactory);
     this.contract = contract;
   }
-
-  /**
-   *  Unpause the module
-   */
-  public unpause = async (params: TxParams): Promise<PolyResponse> => {
-    assert.assert(await this.paused(), ErrorCode.PreconditionRequired, 'Controller not currently paused');
-    assert.assert(
-      await this.isCallerTheSecurityTokenOwner(params.txData),
-      ErrorCode.Unauthorized,
-      'Sender is not owner',
-    );
-    return (await this.contract).unpause.sendTransactionAsync(params.txData, params.safetyFactor);
-  };
-
-  /**
-   *  Check if module paused
-   */
-  public paused = async (): Promise<boolean> => {
-    return (await this.contract).paused.callAsync();
-  };
-
-  /**
-   *  Pause the module
-   */
-  public pause = async (params: TxParams): Promise<PolyResponse> => {
-    assert.assert(!(await this.paused()), ErrorCode.ContractPaused, 'Controller currently paused');
-    assert.assert(
-      await this.isCallerTheSecurityTokenOwner(params.txData),
-      ErrorCode.Unauthorized,
-      'Sender is not owner',
-    );
-    return (await this.contract).pause.sendTransactionAsync(params.txData, params.safetyFactor);
-  };
-
-  /**
-   *The maximum number of concurrent token holders
-   */
-  public maxHolderCount = async (): Promise<number> => {
-    return (await (await this.contract).maxHolderCount.callAsync()).toNumber();
-  };
-
-  /**
-   * Used to verify the transfer transaction and prevent a transfer if it passes the allowed amount of token holders
-   * @return boolean transfer result, address
-   */
-  public verifyTransfer = async (params: VerifyTransferParams): Promise<VerifyTransfer> => {
-    assert.isETHAddressHex('from', params.from);
-    assert.isETHAddressHex('to', params.to);
-    const decimals = await (await this.securityTokenContract()).decimals.callAsync();
-    const result = await (await this.contract).verifyTransfer.callAsync(
-      params.from,
-      params.to,
-      valueToWei(params.amount, decimals),
-      params.data,
-    );
-    const transferResult = parseTransferResult(result[0]);
-    return {
-      transferResult,
-      address: result[1],
-    };
-  };
-
-  /**
-   * Sets the cap for the amount of token holders there can be
-   */
-  public changeHolderCount = async (params: ChangeHolderCountParams): Promise<PolyResponse> => {
-    assert.assert(
-      await this.isCallerAllowed(params.txData, Perm.Admin),
-      ErrorCode.Unauthorized,
-      'Caller is not allowed',
-    );
-    return (await this.contract).changeHolderCount.sendTransactionAsync(
-      numberToBigNumber(params.maxHolderCount),
-      params.txData,
-      params.safetyFactor,
-    );
-  };
 
   /**
    * Subscribe to an event type emitted by the contract.
@@ -226,7 +110,9 @@ export default class CountTransferManagerWrapper extends ModuleWrapper {
    * Gets historical logs without creating a subscription
    * @return Array of logs that match the parameters
    */
-  public getLogsAsync: GetCountTransferManagerLogsAsyncParams = async <ArgsType extends CountTransferManagerEventArgs_3_0_0>(
+  public getLogsAsync: GetCountTransferManagerLogsAsyncParams = async <
+    ArgsType extends CountTransferManagerEventArgs_3_0_0
+  >(
     params: GetLogsAsyncParams,
   ): Promise<LogWithDecodedArgs<ArgsType>[]> => {
     assert.doesBelongToStringEnum('eventName', params.eventName, CountTransferManagerEvents_3_0_0);
@@ -239,4 +125,16 @@ export default class CountTransferManagerWrapper extends ModuleWrapper {
     );
     return logs;
   };
+}
+
+export function isCountTransferManager_3_0_0(
+  wrapper: CountTransferManagerCommon,
+): wrapper is CountTransferManager_3_0_0 {
+  return wrapper.contractVersion === ContractVersion.V3_0_0;
+}
+
+export function isCountTransferManager_3_1_0(
+  wrapper: CountTransferManagerCommon,
+): wrapper is CountTransferManager_3_0_0 {
+  return wrapper.contractVersion === ContractVersion.V3_1_0;
 }
