+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/__tests__/3.0.0.test.ts
@@ -1,0 +1,57 @@
+// CountTransferManager test
+import { mock, instance, reset, when, verify, objectContaining } from 'ts-mockito';
+import {
+  CountTransferManagerContract_3_0_0,
+  ISecurityTokenContract_3_0_0,
+  PolyTokenEvents_3_0_0,
+  BigNumber,
+  Web3Wrapper,
+} from '@polymathnetwork/abi-wrappers';
+import ContractFactory from '../../../../../factories/contractFactory';
+import { CountTransferManager_3_0_0 } from '../3.0.0';
+
+describe('CountTransferManagerWrapper', () => {
+  let target: CountTransferManager_3_0_0;
+  let mockedWrapper: Web3Wrapper;
+  let mockedContract: CountTransferManagerContract_3_0_0;
+  let mockedContractFactory: ContractFactory;
+  let mockedSecurityTokenContract: ISecurityTokenContract_3_0_0;
+
+  beforeAll(() => {
+    mockedWrapper = mock(Web3Wrapper);
+    mockedContract = mock(CountTransferManagerContract_3_0_0);
+    mockedContractFactory = mock(ContractFactory);
+    mockedSecurityTokenContract = mock(ISecurityTokenContract_3_0_0);
+
+    const myContractPromise = Promise.resolve(instance(mockedContract));
+    target = new CountTransferManager_3_0_0(
+      instance(mockedWrapper),
+      myContractPromise,
+      instance(mockedContractFactory),
+    );
+  });
+
+  afterEach(() => {
+    reset(mockedWrapper);
+    reset(mockedContract);
+    reset(mockedContractFactory);
+    reset(mockedSecurityTokenContract);
+  });
+
+  describe('SubscribeAsync', () => {
+    test('should throw as eventName does not belong to CountTransferManager', async () => {
+      // Mocked parameters
+      const mockedParams = {
+        eventName: PolyTokenEvents_3_0_0.Transfer,
+        indexFilterValues: {},
+        callback: () => {},
+        isVerbose: false,
+      };
+
+      // Real call
+      await expect(target.subscribeAsync(mockedParams)).rejects.toEqual(
+        new Error(`Expected eventName to be one of: 'ModifyHolderCount', 'Pause', 'Unpause', encountered: Transfer`),
+      );
+    });
+  });
+});
