+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/index.ts
@@ -1,0 +1,7 @@
+/* istanbul ignore file */
+import { CountTransferManager_3_0_0, isCountTransferManager_3_0_0, isCountTransferManager_3_1_0 } from './3.0.0';
+
+export type CountTransferManager = CountTransferManager_3_0_0;
+
+export { CountTransferManager_3_0_0, isCountTransferManager_3_0_0, isCountTransferManager_3_1_0 };
+export { default as CountTransferManagerCommon } from './common';
