+++ src/factories/moduleWrapperFactory.ts
@@ -3,7 +3,7 @@ import {
   GeneralTransferManager_3_0_0,
   GeneralTransferManager_3_1_0,
 } from '../contract_wrappers/modules/transfer_manager/general_transfer_manager_wrapper';
-import CountTransferManagerWrapper from '../contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper';
+import { CountTransferManager_3_0_0 } from '../contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper';
 import ManualApprovalTransferManagerWrapper from '../contract_wrappers/modules/transfer_manager/manual_approval_transfer_manager_wrapper';
 import PercentageTransferManagerWrapper from '../contract_wrappers/modules/transfer_manager/percentage_transfer_manager_wrapper';
 import LockUpTransferManagerWrapper from '../contract_wrappers/modules/transfer_manager/lock_up_transfer_manager_wrapper';
@@ -83,7 +83,7 @@ interface GetEtherDividendCheckpoint extends GetModuleParams {
 
 interface GetModuleInstance {
   (params: GetGeneralPermissionManager): Promise<GeneralPermissionManagerWrapper>;
-  (params: GetCountTransferManager): Promise<CountTransferManagerWrapper>;
+  (params: GetCountTransferManager): Promise<CountTransferManager_3_0_0>;
   (params: GetGeneralTransferManager): Promise<GeneralTransferManager_3_0_0 | GeneralTransferManager_3_1_0>;
   (params: GetManualApprovalTransferManager): Promise<ManualApprovalTransferManagerWrapper>;
   (params: GetPercentageTransferManager): Promise<PercentageTransferManagerWrapper>;
@@ -132,7 +132,7 @@ export default class ModuleWrapperFactory {
         break;
       // TMs
       case ModuleName.CountTransferManager:
-        moduleWrapper = new CountTransferManagerWrapper(
+        moduleWrapper = new CountTransferManager_3_0_0(
           this.web3Wrapper,
           this.contractFactory.getCountTransferManagerContract(params.address),
           this.contractFactory,
