+++ src/index.ts
@@ -93,7 +93,10 @@ export {
   isGeneralTransferManager_3_1_0,
 } from './contract_wrappers/modules/transfer_manager/general_transfer_manager_wrapper';
 export {
-  default as CountTransferManager,
+  CountTransferManager,
+  CountTransferManager_3_0_0,
+  isCountTransferManager_3_0_0,
+  isCountTransferManager_3_1_0,
 } from './contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper';
 export {
   default as ManualApprovalTransferManager,
